welbexcar/requirements.txt
amqp==5.1.1
asgiref==3.7.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==3.2.19
django-celery-results==2.5.1
djangorestframework==3.14.0
drf-yasg==1.21.5
flake8==5.0.4
geographiclib==2.0
geopy==2.3.0
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.1
pycodestyle==2.9.1
pyflakes==2.5.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
